build: renamed server-middleware-* to server-plugin-*
