+++ .release-please-manifest.json
@@ -1,11 +1,13 @@
 {
     "packages/authup": "1.0.0-beta.9",
-    "packages/config": "1.0.0-beta.9",
-    "packages/core-kit": "1.0.0-beta.9",
     "packages/client-web": "1.0.0-beta.9",
     "packages/client-web-config": "1.0.0-beta.9",
     "packages/client-web-kit": "1.0.0-beta.9",
-    "packages/server-adapter": "1.0.0-beta.9",
+    "packages/config": "1.0.0-beta.9",
+    "packages/core-kit": "1.0.0-beta.9",
     "packages/server-core": "1.0.0-beta.9",
-    "packages/server-kit": "1.0.0-beta.9"
+    "packages/server-kit": "1.0.0-beta.9",
+    "packages/server-plugin-http": "1.0.0-beta.9",
+    "packages/server-plugin-kit": "1.0.0-beta.9",
+    "packages/server-plugin-socket-io": "1.0.0-beta.9"
 }
