+++ packages/server-plugin-http/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "@authup/server-middleware-http",
+    "name": "@authup/server-plugin-http",
     "version": "1.0.0-beta.9",
     "description": "An adapter package to integrate into other services.",
     "exports": {
@@ -38,7 +38,7 @@
     },
     "homepage": "https://github.com/authup/authup#readme",
     "dependencies": {
-        "@authup/server-middleware-kit": "^1.0.0-beta.9"
+        "@authup/server-plugin-kit": "^1.0.0-beta.9"
     },
     "devDependencies": {
         "@authup/core-kit": "^1.0.0-beta.9",
