+++ packages/server-plugin-http/src/index.ts
@@ -6,4 +6,4 @@
  */
 
 export * from './module';
-export * from './type';
+export * from './types';
