+++ packages/server-plugin-socket-io/.editorconfig
