+++ packages/server-plugin-socket-io/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "@authup/server-middleware-kit",
+    "name": "@authup/server-middleware-socket-io",
     "version": "1.0.0-beta.9",
-    "description": "A package containing utilities for middleware solutions.",
+    "description": "An adapter package to integrate into other services.",
     "exports": {
         "./package.json": "./package.json",
         ".": {
@@ -19,23 +19,9 @@
     "scripts": {
         "build:types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
         "build:js": "rollup -c",
-        "build": "rimraf ./dist && cross-env NODE_ENV=production npm run build:js && npm run build:types",
-        "test": "cross-env NODE_TLS_REJECT_UNAUTHORIZED=0 NODE_ENV=test jest --config ./test/jest.config.js",
-        "test:coverage": "npm run test -- --coverage"
+        "build": "rimraf ./dist && cross-env NODE_ENV=production npm run build:js && npm run build:types"
     },
-    "keywords": [
-        "auth",
-        "module",
-        "typescript",
-        "middleware",
-        "http",
-        "express",
-        "socket",
-        "socket.io",
-        "websocket",
-        "adapter",
-        "oauth2"
-    ],
+    "keywords": [],
     "author": {
         "name": "Peter Placzek",
         "email": "contact@tada5hi.net",
@@ -45,20 +31,18 @@
     "repository": {
         "type": "git",
         "url": "git+https://github.com/authup/authup.git",
-        "directory": "server/packages/adapter"
+        "directory": "packages/server-middleware-http"
     },
     "bugs": {
         "url": "https://github.com/authup/authup/issues"
     },
     "homepage": "https://github.com/authup/authup#readme",
     "dependencies": {
-        "jose": "^5.2.4",
-        "redis-extension": "^1.3.0"
+        "@authup/server-plugin-kit": "^1.0.0-beta.9"
     },
     "devDependencies": {
         "@authup/core-kit": "^1.0.0-beta.9",
-        "@authup/server-kit": "^1.0.0-beta.9",
-        "@hapic/oauth2": "^2.4.2"
+        "@authup/server-kit": "^1.0.0-beta.9"
     },
     "peerDependencies": {
         "@authup/core-kit": "^1.0.0-beta.9",
